requirements-prod.txt
# Production-Only Requirements (Minimal for Docker)
# Excludes: pytest, ruff, pip-audit, opentelemetry (dev/monitoring tools)

# Web Framework
fastapi~=0.128.0
uvicorn[standard]~=0.40.0
websockets~=15.0.1

# Database
sqlalchemy~=2.0.45
asyncpg~=0.31.0
alembic~=1.17.2

# AI/ML Services
azure-cognitiveservices-speech>=1.34.0
groq>=1.0.0

# HTTP Client
httpx~=0.28.1

# Configuration
python-dotenv~=1.2.1
pydantic-settings~=2.12.0

# Forms
python-multipart~=0.0.21

# Templates
jinja2~=3.1.6

# Twilio - REMOVED: Not used in production (only Telnyx)
# twilio~=10.4.0.2

# Rate Limiting
slowapi~=0.1.9

# Webhook Security - CRITICAL: Use stable versions
cryptography>=42.0.0,<45.0.0
h11>=0.16.0
urllib3>=2.6.3
httpcore>=1.0.9

# Input Sanitization
markupsafe~=2.1.5

# State Management
redis[hiredis]~=5.0.1

# Audio processing
numpy>=1.26.0

# Monitoring (Minimal)
prometheus-client>=0.20.0

# Logging
python-json-logger~=2.0.7
asgi-correlation-id~=4.3.1
